+++ src/main/java/com/sanket/contactdirectory/dao/PersonRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import com.sanket.contactdirectory.entity.Person;
 
 @Repository
-public interface PersonRepository extends JpaRepository<Person, Long> {
+public interface PersonRepository extends JpaRepository<Person, Integer> {
 
 }
