+++ src/main/java/com/sanket/contactdirectory/service/ContactService.java
@@ -8,11 +8,11 @@ public interface ContactService {
 
     List<Contact> getAllContacts();
 
-    Contact getContactById(long id);
+    Contact getContactById(int id);
 
     Contact addContact(Contact contact);
 
-    Contact updateContact(long id, Contact contact);
+    Contact updateContact(int id, Contact contact);
 
-    String deleteContact(long id);
+    String deleteContact(int id);
 }
